collector/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
clickhouse-driver==0.2.6
pyyaml==6.0.1
numpy==1.24.3
